fixing components
